Support big endian platforms
- KTX file headers can be stored in either little or big endian mode.
Check the endianness field of the header and swap headers if necessary.
- For DDS files, read the magic word 'DDS ' as a uint32_t to determine
if we need to swap headers. Since DDS files should always be stored in
little endian byte order, this should only occur on big endian
platforms. While saving DDS files, check if the system is big endian and
if so, swap headers again before writing the header.